requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
dash==4.1.0
starlette==1.0.0
dash-bootstrap-components==2.0.4
dash-cytoscape==1.0.2
dash-extensions==2.0.5
alembic==1.18.4
sqlalchemy==2.0.49
asyncpg==0.31.0
pydantic-settings==2.13.1
psycopg2-binary==2.9.11
openai==2.32.0
tiktoken==0.12.0
python-dotenv==1.2.2
cryptography==46.0.7
langchain==1.2.15
langchain-openai==1.1.14
langchain-community==0.4.1
langchain-neo4j==0.9.0
requests==2.33.1
networkx==3.6.1
python-louvain==0.16
mcp[cli]==1.27.0
